fix an issue where regression/394.lox was failing.
By switching to Stmt.Class super class from Expr.Variable to
Expr.Expr fixed the issue. I *think* it was due to
locals: std.AutoHashMap(*const Expr.Expr, usize)

Because it is hashing on the pointers, previously I think it was
working due to the same pointer being used, where as in this case
it was a copy of the data and hashed differently?